+++ src/layers/DataBlockLayer.js
@@ -3,9 +3,9 @@
 		//this layer outputs the inputs with no changes
 		constructor(size) {
 			this.nextLayer; //the connected layer
-			this.inData = new Float32Array(size); //the inData
-			this.outData = new Float32Array(size); //will be init when "connect" is called.
-			this.costs = new Float32Array(size); //costs for each neuron
+			this.inData = new Float64Array(size); //the inData
+			this.outData = new Float64Array(size); //will be init when "connect" is called.
+			this.costs = new Float64Array(size); //costs for each neuron
 			this.pl; //reference to previous layer
 		}
 
@@ -16,9 +16,9 @@
 			// try to connect to it
 			if (this.inData.length == 0) {
 				//if not already initialized
-				this.inData = new Float32Array(layer.outSize());
-				this.outData = new Float32Array(layer.outSize());
-				this.costs = new Float32Array(layer.outSize());
+				this.inData = new Float64Array(layer.outSize());
+				this.outData = new Float64Array(layer.outSize());
+				this.costs = new Float64Array(layer.outSize());
 			}
 			this.pl = layer;
 		}
