+++ src/layers/ResReceiverLayer.js
@@ -20,8 +20,8 @@
 		}
 
 		set previousLayer(layer) {
-			this.inData = new Float32Array(layer.outSize());
-			this.costs = new Float32Array(layer.outSize());
+			this.inData = new Float64Array(layer.outSize());
+			this.costs = new Float64Array(layer.outSize());
 			this.pl = layer;
 			//time to find this layers soulmate
 			let found = false;
@@ -43,11 +43,11 @@
 				if (layer.outSize() != this.emitter.outSize()) {
 					throw "emitter size must equal the size of the previous layer of the corresponding receiver layer";
 				}
-				this.outData = new Float32Array(layer.outSize());
+				this.outData = new Float64Array(layer.outSize());
 			} else if (this.mode == "concat") {
-				this.outData = new Float32Array(layer.outSize() + this.emitter.outSize());
+				this.outData = new Float64Array(layer.outSize() + this.emitter.outSize());
 			}
-			this.costsForEmitter = new Float32Array(this.emitter.outSize());
+			this.costsForEmitter = new Float64Array(this.emitter.outSize());
 		}
 
 		//we dont look in front because residual data only goes forwards.
